+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - highlight the first unread message upon opening a chat #4525
 - copy files to internal tmp dir before opening attachements #4498
 - enable notifications on mentions in muted chats #4538
+- e2e test for group creation #4614
 - always show accounts with unread messages, even when they're normally scrolled out of view #4536
 - make log additionally available at "Settings / Advanced / View Log" #4610